06th Place/requirements.txt
absl-py==0.7.0
alabaster==0.7.12
appnope==0.1.0
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
awscli==1.16.88
Babel==2.6.0
backcall==0.1.0
bleach==3.1.0
botocore==1.12.78
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
colorama==0.3.9
coverage==4.5.2
cycler==0.10.0
dask==1.1.1
decorator==4.3.0
defusedxml==0.5.0
distributed==1.25.3
docopt==0.6.2
docutils==0.14
entrypoints==0.3
flake8==3.6.0
future==0.17.1
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
HeapDict==1.0.0
idna==2.8
imagesize==1.1.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jmespath==0.9.3
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lightgbm==2.2.2
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==5.0.0
msgpack==0.6.1
munch==2.3.2
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numexpr==2.6.9
numpy==1.15.4
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.1
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.8.1
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.6.0
py==1.7.0
py-cpuinfo==4.0.0
pyasn1==0.4.5
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.3.1
pyparsing==2.3.0
pytest==4.1.1
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.9
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
rsa==3.4.2
s3transfer==0.1.13
sacred==0.7.4
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==1.2.1
sortedcontainers==2.1.0
Sphinx==1.8.3
sphinxcontrib-websupport==1.1.0
# -e git+git@bitbucket.org:walen/drivendata-cleaning-time.git@5c5ffd51c5d1595477a34b270e62d4163e51e2ea#egg=src
statsmodels==0.9.0
tables==3.4.4
tblib==1.3.2
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.29.1
traitlets==4.3.2
tsfresh==0.11.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.1
zict==0.1.3
joblib
logger